+++ src/ts/ir/expandMarker.ts
@@ -49,7 +49,6 @@ export const expandMarker = (range: Range, vditor: IVditor) => {
         item.classList.remove("vditor-ir__node--expand");
     });
 
-
     const nodeElement = hasTopClosestByClassName(range.startContainer, "vditor-ir__node");
     const nodeElementEnd = !range.collapsed && hasTopClosestByClassName(range.endContainer, "vditor-ir__node");
     // 选中文本为同一个 nodeElement 内时，需要展开
@@ -57,7 +56,6 @@ export const expandMarker = (range: Range, vditor: IVditor) => {
         return;
     }
 
-    
     if (nodeElement) {
         nodeElement.classList.add("vditor-ir__node--expand");
         nodeElement.classList.remove("vditor-ir__node--hidden");