+++ demo/index.js
@@ -50,7 +50,7 @@ if (window.innerWidth < 768) {
 }
 
 window.vditor = new Vditor('vditor', {
-  _lutePath: `http://192.168.80.35:9090/lute.min.js?${new Date().getTime()}`,
+  _lutePath: `http://192.168.0.107:9090/lute.min.js?${new Date().getTime()}`,
   // _lutePath: 'src/js/lute/lute.min.js',
   toolbar,
   mode: 'ir',