+++ demo/render.js
@@ -58,12 +58,14 @@ window.setTheme = (theme) => {
   const outlineElement = document.getElementById('outline')
   if (theme === 'dark') {
     Vditor.setCodeTheme('native')
-    Vditor.setContentTheme('dark', 'https://cdn.jsdelivr.net/npm/vditor/dist/css/content-theme')
+    Vditor.setContentTheme('dark',
+      'https://cdn.jsdelivr.net/npm/vditor/dist/css/content-theme')
     outlineElement.classList.add('dark')
     document.querySelector('html').style.backgroundColor = '#2f363d'
   } else {
     Vditor.setCodeTheme('github')
-    Vditor.setContentTheme('light', 'https://cdn.jsdelivr.net/npm/vditor/dist/css/content-theme')
+    Vditor.setContentTheme('light',
+      'https://cdn.jsdelivr.net/npm/vditor/dist/css/content-theme')
     outlineElement.classList.remove('dark')
     document.querySelector('html').style.backgroundColor = '#fff'
   }