+++ CHANGELOG.md
@@ -79,6 +79,7 @@
 
 ### v3.4.5 / 2020-08-xx
 
+* [694](https://github.com/Vanessa219/vditor/pull/694) Fix expandmarker for node element `修复缺陷`
 * [697](https://github.com/Vanessa219/vditor/pull/697) 修复Lute类型的定义 `开发重构`
 
 ### v3.4.4 / 2020-08-05